refactor: Adicionando novo background e organizando estrutura de pastas.
