+++ frontend/src/components/animation/gradualSpacing.tsx
