+++ frontend/src/components/animation/hooks/isVisible.tsx
