+++ frontend/src/pages/home/Home.tsx
@@ -1,12 +1,11 @@
-import Banner from './components/banner/Banner';
-import Content from './components/content/Content';
+import Banner from './banner/Banner';
+import Content from './content/Content';
 
 export default function Home () {
     return (
       <div>
         <Banner />
         <Content />
-        
       </div>
     )
 }
